fix: Ensure grid is built and robots set course to a destination
Previously, robots didn't even know how to enter the grid, but you
can't blame them since they had no means of realizing the
possibility of the grid. And that's okay because up until now, the
creator (engine) had also not realized the grid. This commit
provides the creator with enough information to generate the grid,
as well where to place robots on the grid. Methods were also
introduced to cater to the robots' need of moving around the grid
and relating experience to one another. Now they actually have a
shot at arriving at their desired destinations.

Also, to future proof and increase code readibility, I upgraded us
to ES6/7.